feat: Add share settings functionality
Adds a 'Share Settings' button to the settings menu. Clicking this menu
will copy a share link to the clipboard. This share link is encoded with
all of the current settings. Clicking on or navigating to a share link
will cause all of the settings from that link to be set.

This feature is useful for sharing specific configurations with
different groups of people. For example, a teacher might want to post a
link to similar rhythms for younger students, and more complicated
rhythms for older students.

Another useful way to use this feature is to bookmark share links. This
allows users to quickly get back to a specific configuration.